requirement.txt
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
fake-headers==1.0.2
fake-useragent==1.1.3
greenlet==2.0.2
html5lib==1.1
idna==3.4
pycparser==2.21
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.41
urllib3==1.26.15
webencodings==0.5.1
